deactivating PVT and GroundCollector for MVP. giving up on merger for solar tech for MVP, let's use the honey-badger for now
